requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
deepgram-sdk==0.2.5
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
python-dotenv==0.19.2
websockets==10.2
yarl==1.7.2
